requirements.txt
google-generativeai
python-dotenv
yagmail
requests
